Move confirmationModal to its own component directory
