+++ src/components/BackOffice/CloseWorkOrders/index.js
@@ -4,7 +4,7 @@ import ErrorMessage from '../../Errors/ErrorMessage'
 import { Button, TextArea, TextInput } from '../../Form'
 import DatePicker from '../../Form/DatePicker'
 import Spinner from '../../Spinner'
-import ConfirmationModal from '../Components/ConfirmationModal'
+import ConfirmationModal from '../../ConfirmationModal'
 import ControlledRadio from '../Components/ControlledRadio'
 import SuccessMessage from '../Components/SuccessMessage'
 import Layout from '../Layout'
